A translation imporvement suggestion
The translation located in line 10 has a tiny imporvement could be done. The phrase 'a useful skill by itself' should be considered as a whole meaning 「自身就是一项有用的技能」rather than split it up, take the 'by itself' part sticking it to 'learning to program'. That would be a diffrent meaning. Beside, it makes more sense translating this way considering the sentence right after it. The author made a comparision about programming serves as an useful skill itself alone and in the situations it serves as way to solve other problems.

Other modifications are trival and could be ignored since in most case they are just a matter of taste and would result in no misunderstandings.